- change: revert to original tray right-click dialog and adapted it for 0.1 increase (Stulle)

+++ srchybrid/emuleDlg.cpp
@@ -2893,8 +2893,8 @@ void CemuleDlg::OnTrayRButtonUp(CPoint pt)
 										thePrefs.GetMaxGraphUploadRate(true), thePrefs.GetMaxGraphDownloadRate(),
 										thePrefs.GetMaxUpload(), thePrefs.GetMaxDownload());
 										*/
-										(int)thePrefs.GetMaxGraphUploadRate(), (int)thePrefs.GetMaxGraphDownloadRate(),
-										(int)thePrefs.GetMaxUpload(), (int)thePrefs.GetMaxDownload());
+										thePrefs.GetMaxGraphUploadRate(), thePrefs.GetMaxGraphDownloadRate(),
+										thePrefs.GetMaxUpload(), thePrefs.GetMaxDownload());
 										//Xman end
 	if (m_pSystrayDlg)
 	{
